Refactor: Move the metrics webserver into metrics/
Keeps the main package pretty small to just flag processing and setup.
