+++ metrics/metrics_server_test.go
@@ -1,4 +1,4 @@
-package main
+package metrics
 
 import "testing"
